Added bonuses to Style System
The style system will now be accompanied with ULTRAKILL-esque style bonuses in the top left (or top centre, depending on your settings) of the screen.
I was originally going to find some convoluted way to display the bonuses under the style meter, but not only would that get in the way a lot, it'd also just be a pain to implement. This is way easier.