The player can now Enemy Jump
An Enemy Jump is a Wall Jump performed on an enemy. This technique is similar to Enemy Step from the Devil May Cry series, but rather than facilitating Jump Cancelling, Enemy Jump not only allows the player to gain distance from the enemy, but also recharges their double jump (if applicable), like the regular Wall Jump.